% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dar_semos.R
\name{dar_semos}
\alias{dar_semos}
\title{Deseasonalized Autoregressive smooth EMOS (DAR-SEMOS)}
\usage{
dar_semos(
  train,
  test,
  doy_col,
  obs_col,
  mean_col,
  sd_col,
  n_ahead = 0,
  aic = TRUE,
  order.max = NULL,
  ...
)
}
\arguments{
\item{train}{data frame containing the training data.}

\item{test}{data frame containing the testing data.}

\item{doy_col}{column of the variable day of the year.}

\item{obs_col}{column of the observation variable.}

\item{mean_col}{column of the ensemble mean forecast variable.}

\item{sd_col}{column of the ensemble standard deviation variable.}

\item{n_ahead}{integer corresponding to the forecast ahead time
(0 for ahead times not greater than 24 hours,
1 for ahead times greater than 24 hours and not greater than 48 hours, and so on).}

\item{aic}{logical; if \code{TRUE} (default), then the Akaike Information Criterion
is used to choose the order of the autoregressive model.
If \code{FALSE}, the model of order \code{order.max} is fitted.}

\item{order.max}{maximum order (or order) of model to fit.}

\item{...}{unused}
}
\value{
A data frame containing the distribution (location and scale) parameters.
}
\description{
Deseasonalized autoregressive smooth EMOS (DAR-SEMOS) as described by Jobst, Möller and Groß (2024).
}
\examples{
# load data for station Hannover
data(station)

# select data for lead time 24 hours
data <- station[station$lt == 24, ]

# split data in training and test data
train <- data[data$date <= as.Date("2019-12-31"), ]
test <- data[data$date > as.Date("2019-12-31"), ]

# estimate model
fit <- dar_semos(train = train,
                 test = test,
                 doy_col = 3,
                 obs_col = 9,
                 mean_col = 10,
                 sd_col = 11,
                 n_ahead = 0)

# distribution parameters
head(fit)

# select data for lead time 48 hours
data <- station[station$lt == 48, ]

# split data in training and test data
train <- data[data$date <= as.Date("2019-12-31"), ]
test <- data[data$date > as.Date("2019-12-31"), ]

# estimate model
fit <- dar_semos(train = train,
                 test = test,
                 doy_col = 3,
                 obs_col = 9,
                 mean_col = 10,
                 sd_col = 11,
                 n_ahead = 1)

# distribution parameters
head(fit)

}
\references{
Jobst, D., Möller, A., and Groß, J. (2024). Time Series based Ensemble Model Output Statistics for Temperature Forecasts Postprocessing.
}
