% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/semos.R
\name{semos}
\alias{semos}
\title{Smooth EMOS (SEMOS)}
\usage{
semos(train, test, doy_col, obs_col, mean_col, sd_col, ...)
}
\arguments{
\item{train}{data frame containing the training data.}

\item{test}{data frame containing the testing data.}

\item{doy_col}{column of the variable day of the year.}

\item{obs_col}{column of the observation variable.}

\item{mean_col}{column of the ensemble mean forecast variable.}

\item{sd_col}{column of the ensemble standard deviation variable.}

\item{...}{unused}
}
\value{
A data frame containing the distribution (location and scale) parameters.
}
\description{
Smooth EMOS (SEMOS) as described by Jobst, Möller and Groß (2024).
}
\examples{
# load data for station Hannover
data(station)

# select data for lead time 24 hours
data <- station[station$lt == 24, ]

# split data in training and test data
train <- data[data$date <= as.Date("2019-12-31"), ]
test <- data[data$date > as.Date("2019-12-31"), ]

# estimate model
fit <- semos(train = train,
             test = test,
             doy_col = 3,
             obs_col = 9,
             mean_col = 10,
             sd_col = 11)

# distribution parameters
head(fit)

# select data for lead time 48 hours
data <- station[station$lt == 48, ]

# split data in training and test data
train <- data[data$date <= as.Date("2019-12-31"), ]
test <- data[data$date > as.Date("2019-12-31"), ]

# estimate model
fit <- semos(train = train,
             test = test,
             doy_col = 3,
             obs_col = 9,
             mean_col = 10,
             sd_col = 11)

# distribution parameters
head(fit)

}
\references{
Jobst, D., Möller, A., and Groß, J. (2024). Time Series based Ensemble Model Output Statistics for Temperature Forecasts Postprocessing. https://doi.org/10.48550/arXiv.2402.00555.
}
